readyset-mysql: Add sleeps to fallback tests
Add a sleep between the creation of the table and queries that depend on
those tables existing to all tests in the readyset-mysql fallback test
suite, to allow a little bit of time for the DDL to replicate. Without
this, this test fails occasionally in CI now that we're running things
in parallel (I assume because things are better isolated, but I'm not
entirely sure why that caused this to start happening).

Change-Id: I81c0e0cef6d9a33e98953ef959df10cbccd288d0
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3217
Tested-by: Buildkite CI
Reviewed-by: Peter Farr <peter@readyset.io>