replicators: Send anonymized schemas to segment
When we replicate, we perform queries to find the create statements for
the tables and views of a database. This change collects those schemas
into some helper structs, anonymizes them, and sends them to segment if
a snapshot was taken.

This adds some extra string clones in the snapshot paths, but only
for the create statements, which should be small compared to the overall
time taken by replication.

Change-Id: I013d94457419a2f90053270056e5f913db2d731b
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3186
Reviewed-by: Alex Martin <Alex.m@readyset.io>
Tested-by: Buildkite CI