dataflow: Don't record remapped keys in full materializations
If we remap a downstream key into an upstream key in a full
materialization, we're *guaranteed* never to get any evictions from
those upstream keys (since full materializations can't evict), so
recording the remap is pointless - and in fact has the potential to
unboundedly increase memory! Instead, since the miss is definitely
going to be a node in the current domain, we can just check to see if
it's partial, and only insert the remapped keys if it is.

Change-Id: I33155664630dccae0c6acc4a12503da8a70596d7
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3222
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Tested-by: Buildkite CI