readyset-client: Reject adhoc MySQL queries that include placeholders
The code paths for prepared statements and adhoc queries share a lot of
common code, and it turns out that this causes issues if the client
sends an adhoc query that includes placeholders for parameters.

Normally even for adhoc queries we replace the parameters with
placeholders in rewrite::process_query, and then return the actual
parameter values as a separate list. Later code then expects the number
of placeholders in the query to match the length of the parameter list.
However, if there are already literal placeholder markers included in
the text of the query, then this expectation will be violated, and
strange things will happen. In some cases with a single placeholder we
hit the `if raw_keys.is_empty()` conditional in `build_view_query`, and
return `bogo` which has a value of 0, so the query ends up behaving as
though we had specified 0 for the parameter. In other cases, we do not
hit that conditional, and the worker thread panics.

To fix this, this commit adds a check in the adhoc query code path
before rewriting the statement, and return an error if any placeholder
literals are found.

Change-Id: I003d9d32dbb9942d97ff63042213df68f0f3dcb3
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3224
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>