telemetry: Add test reporter and unit tests
This adds a TestTelemetryReporter that intercepts the events and
payloads sent by the TelemetrySender, does not send them to Segment, and
allows tests to inspect and assert on what events and payloads were sent
and how many times they were sent.

Additionally, 2 unit tests were added and the incorrect behaviors
(sending CreateCache isntead of ShowCaches and ShowProxiedQueries) were
fixed.

Change-Id: I3a3d5c87f00b8e617c352e24282ed2144edda662
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3225
Tested-by: Buildkite CI
Reviewed-by: Alex Martin <Alex.m@readyset.io>