replicators: support versions of Postgres pre 14
We previously supported versions 13 and older of Postgres by
replicating DDL using a dedicated table that we stored DDL
statements into, and they were replicated using the regular WAL
messages. Then we switched to using `pg_logical_emit_message`
instead, which apparently only works with for version 14 and up.

This change reintroduces support for older versions of Postgres
using the dedicated table method. Unlike with the previous
implementation the new table only ever contains exactly one row
and this row holds the latest DDL statement, and the updates to
this row are replicated using the WAL update message. Postgres
14 and up will keep using `pg_logical_emit_message`.

Change-Id: I1235012b038476df13849d4e9def64d093449d1c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3227
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>