telemetry: Add TelemetryEvent for proxied queries
This adds a new TelemetryEvent for proxied queries, which will be sent
as an anonymized DeniedQuery.

In order for DeniedQuery and its fields to be available to the telemetry
reporter, this lifts out those types from query_status_cache.rs to a new
module -- readyset/query.rs.

Additionally, in order to correctly anonymize the full query, the
implementation of Query::to_anonymized_string() was updated to use the
new Anonymizer, or fully anonymize the string if it is a query that
failed to parse.

The TelemetryEvent is as of yet unused--hooking it up to the
TelemetryReporter will be future work.

Change-Id: I6c47b549ce844a44f9c4691ab94e9caaec7061b4
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3237
Tested-by: Buildkite CI
Reviewed-by: Alex Martin <Alex.m@readyset.io>