Add real health reporting to the server
Previously we used the /prometheus endpoint as a proxy for server
health. This commit adds a /health endpoint, and reports using the
health reporter crate. This also adds failpoint to check what health we
report if the controller or worker fail to come up, and adds testing for
each.

This commit also generalizes some testing helpers so they can be used
for both server and adapter testing.

Change-Id: I5aee853befce000f153d940e3927f528b22b76f9
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3259
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke.o@readyset.io>