data: Add a `collation` field to DfType::Text
As the first step towards laying the groundwork for collation support,
this commit defines a new `Collation` enum in readyset-data, which
currently has a single variant representing the collation we currently
implement, UTF-8, and adds a field of that type to `DfType::Text`,
defaulting to the UTF-8 collation everywhere we have to make a
decision.

Since to return this collation the `output_col_type` of
`GroupedOperation` now has to construct a non-constant value, this also
makes that returned an *owned* `DfType` rather than a reference - this
also allows us later to make decisions such as which collation to return
from eg `GROUP_CONCAT`, which will become important down the road.

Change-Id: Ie157f9a43348ceadb5ee244486a64ccbdfed3697
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3264
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>