telemetry: Add fallback anonymous user_id
This adds a fallback user_id in the case that the RS_API_KEY is not set.

The user_id for telemetry reports is now:
1. If the RS_API_KEY is set, that becomes the user_id.
2. Else we attempt to get the machine-id, a unique identifier, and then
use a cryptographically secure hashing library to anonymize it (as the
machine id itself should be considered confidential).
3. If 2 fails (which it may if the machine-uid crate doesn't support the
system it's running on, we fall back to not having any user id.

Change-Id: Iabfaab68f022c9ec9d7a6a730ec148f4b954a38e
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3267
Tested-by: Buildkite CI
Reviewed-by: Alex Martin <Alex.m@readyset.io>