test: Fix key_count_rpc test races via the eventually! macro
This test occasionally fails, and it seems like the failure reason is
most likely due to a possible race condition where we check the key
counts before the write is done being processed. By wrapping the second
check in an eventually! macro we will retry a number of times if the
results don't match up with what the expect at first.

I also moved the location of the initial call to `let key_counts = ...`
to improve clarity.

Change-Id: I0230278db9c594b7a10a58d65546124775eefec8
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3268
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>