data: Add a nibble-sized collation to TinyText
Since the `len` field of TinyText is guaranteed never to be larger than
14 (since it's bounded by the size of the actual data we can store
inside TinyText) we actually have a spare 4 bits inside of TinyText to
store whatever data we want. In this case, we're using it to squeeze in
up to 16 of the most common collations (though right now we're still
just limited to Utf8).

This implements the actual storage and retrieval of collation in
TinyText, but it's left unused for now (also, other Text variants don't
have it).

Refs: ENG-1856
Change-Id: I8c129b13ae86f103d144f03cc119aa3a42db9208
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3271
Tested-by: Buildkite CI
Reviewed-by: Vlad Krasnov <vlad@readyset.io>