Move failpoint to server's HTTP router
Previously we handled failpoint registration as rpc requests to the
controller. When starting a readyset-server, the controller is started
after servers http router is started up. By moving the failpoint
registration to the server's http router, we can test what happens when
the rest of the system fails to come up (worker, controller, etc.).

This commit moves the failpoint registration to the servers http router
to accomplish that end.

Change-Id: Ie8e985b31f05de2c221fe6ecf0a54cd6ce15c858
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3272
Tested-by: Buildkite CI
Reviewed-by: Sif Hall <sif@readyset.io>