data: Add a collation to Text
Add a collation field to the header of Text, to go along with the
collation we squeezed into TinyText. Since this is effectively immutable
post construction (since it's behind an Arc), this also comes along with
a few new constructors for building a Text with a collation.

As before, this is currently unused and ignored; actually using it is
coming up soon.

Refs: ENG-1856
Change-Id: I068354f9361add4a651e0ff8265da1849818dafe
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3276
Tested-by: Buildkite CI
Reviewed-by: Vlad Krasnov <vlad@readyset.io>