data: Use collations to impl Hash, PartialEq, and PartialOrd
Delegate to the collation of the string value in the branches of the
Hash, PartialEq, and PartialOrd impls of DfValue to actually decide how
to hash or compare the strings. Since currently the only collation we
implement is UTF8, which is equivalent to the default behavior of
strings, this currently just delegates to the underlying Hash and Ord
impls of the string itself.

Change-Id: I875920d27ea571296307fe6b126ff1d1a78e50d1
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3280
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>