nom-sql: Parse citext
Add a new `SqlType::Citext` variant, parse it, and implement conversion
from it to `DfType::Text` with the Citext collation.

Note that because we're generally not trying to validate that queries
are correct for the dialect, I'm *not* trying to only parse `citext` as
a type if the dialect is PostgreSQL - this leaves things a lot simpler.

Fixes: ENG-1855
Change-Id: Id370b353db3123971f1cf4e8f2d773406044778e
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3285
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>