dataflow: Rename KeyType to PointKey, move to `dataflow-state`
Similar to the old naming of `DataType` (now `DfValue`), the name of the
`KeyType` struct always falsely indicated that it was a "type" of key,
rather than an actual key value itself. The name `PointKey` is much
clearer, and provides a nice parallel with the also-existent `RangeKey`.
Also, since this type is really important to dataflow *state* due to
being integral to the APIs defined there, it makes a lot more sense to
have it defined within the dataflow-state crate rather than in
readyset-common.

Other than and some extra docs that nothing functional is changing here,
but this prepares for putting a *little bit* more logic inside of
`PointKey`.

Refs: ENG-1870
Change-Id: I62a09bd310a12077e0bb2d81eaca2bfe9c200ae6
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3295
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Reviewed-by: Vlad Krasnov <vlad@readyset.io>