readyset-data: Fix panics on overflow and division by zero
Prior to this commit, it was very easy to trigger a server panic by
running statements like `SELECT c FROM t WHERE c = 1 - 2` or `SELECT c /
0 FROM t`. This was because we were using primitive operators to
implement Add/Sub/Mul/Div on DfValue, and for many primitive types these
operators panic when they cause overflow or division by zero.

To fix this, we now use the builtin checked_add/sub/mul/div methods
which return None instead of panicking, and this gets naturally
converted to DfValue::None. Note that this does not necessarily result
in the same behavior as MySQL/Postgres in the applicable conditions, but
it does at least prevent us from panicking.  Also note that the f32/f64
types don't panic, and instead return special values like Inf, so there
are no checked_ methods for those types, and so we fall back on the
primitive operators for f32/f64.

This commit also adds some unit tests. I considered adding logictests as
well but decided not to since the behavior in ReadySet often does not
actually match that of MySQL/Postgres in these situations, and it felt
like a bad idea to add a test that validates behavior that we may not
actually want in the long term.

Change-Id: Ie0d8f590219df91720605b3738d2aef2f3aedcfb
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3301
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>