Support Postgres's UDT in fallback mode
Postgres supports user-defined types, which RS doesn't support
natively. In addition, replication/snapshotting couldn't handle
them, resulting in lots of breakage.

This commit does 4 things:
  * Snapshotting will skip tables with UDTs
  * Replication will also skip tables with UDTs
  * Added enough support in the adapter to be able to leverage the
    fallback connection to upstream, giving us support for uncached
    UDTs

As this is built upon PassThrough, this includes its same
limitations (namely requiring binary transfer mode).

In the nom-sql code, the partial support for ""char"" (different from
the "char" type -- not at all confusing naming!) is hacked around by
explicitly filtering out that type (ENG-1910 was created for that
issue).

This also removes the TableFilter logic from PostgresWalConnector,
as we don't need to check whether the table is filtered or not when
processing WAL events and encounter an unsupported type error, as
we now have the same behavior when filtered and unfiltered (the
majority of the TableFilter logic is still being used in
PostgresReplicator).

Refs: ENG-1823
Change-Id: I6079e4ac4df563639397af455023a9fc3ee4c549
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3307
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>
Reviewed-by: Luke Osborne <luke.o@readyset.io>