dataflow-state: Serialize RangeKey endpoints via PointKey
To get the same normalize-on-serialize behavior as `PointKey`, translate
the bound endpoints of `RangeKey`s to `PointKey`s (which we can do
trivially now that they have the same internal representation) before
serializing. This has the effect of making range lookups on citext work,
so this commit also comes with some new queries added to the citext
logictest to cover those.

Fixes: ENG-1883
Change-Id: Ie8ae7998dcd22a13eaa80df3bedac8ffe1bbe150
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3316
Tested-by: Buildkite CI
Reviewed-by: Vlad Krasnov <vlad@readyset.io>