adapter: move query logger to dedicated thread
Currently the query logger consumes outrageous amounts
of CPU, for example when enabled in the simple_select
system benchmark it consumes 13% of the total CPU used.

This change moves the logger from the shared tokio runtime
to a dedicated single theaded runtime, it still uses the
same amount of CPU, but now it does not interfere with
query execution as much.

Change-Id: Icfedd793149331f32dbac387177a877c5c86c626
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3332
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke.o@readyset.io>