dataflow-expression: Implement GREATEST function
Implement the GREATEST function for both MySQL and PostgreSQL. The type
inference rules for this when comparing values of different types are
remarkably tricky, so most of the complexity in this CL lies there:

    localhost/noria=# select greatest('123', 23);
    greatest
    ----------
        123
    (1 row)

    MySQL [(none)]> select greatest('123', 23);
    +---------------------+
    | greatest('123', 23) |
    +---------------------+
    | 23                  |
    +---------------------+
    1 row in set (0.000 sec)

I've implemented *most* of the rules for how that all works (including a
new `pg_type_category()` method on DfType, which I'm expecting us to be
using quite a bit in the future), but there are a couple of edge cases I
decided not to cover here - I've opened ENG-1909 and ENG-1911 to cover
those edge cases, and there's also comments inline calling out when we
do the wrong thing. See those inline comments too for links to the MySQL
and PostgreSQL docs for how this all works.

Fixes: ENG-1892
Change-Id: Ie10393cd03be452119e3dc5931bcbc6bf22050ee
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3363
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>