dataflow-expression: Implement LEAST function
Since the type inference rules for GREATEST are the trickiest part, and
since the type inference rules for LEAST are the same in both PostgreSQL
and MySQL, I figured I'd go ahead and add support for LEAST in addition
to GREATEST while I was at it. This commit shares as much code as
possible between the two implementations, both in lowering (incl type
inference) and evaluation.

Change-Id: I6b8103470e6a4626266105d2e5109047ccf2e444
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3364
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>