replicators: reused Config for upstream dbs
Currently we have two separate concepts: one is the replication
database, the other one is the upstream database. Both support
the same syntax for url and both require similar additional
configuration flags such as `ssl_root_cert` and
`disable_upstream_ssl_verification`. They are however mutually
exclusive. The server uses the replication url, and the adapter
is using the upstream url, however in embedded or standalone modes
those are one and the same. Point is we have a duplication of flags
that do the same thing, and can be reused. This change reuses
the replication config for upstream connections.

Change-Id: Ibd576d0cfbf015f89ee4c9e7af3c098fce43b553
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3366
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>