replicators: Replicate CITEXT from postgres
Since CITEXT isn't one of the "built-in" types that rust-postgres knows
how to handle, we can't build the type using `Type::from_oid` - instead,
we need to explicitly fetch all the relevant metadata about the type
from the pg catalog, and construct the type ourselves using `Type::new`.
This is slightly more involved, but has the advantage that it paves the
way nicely for custom enum types - there's an `unsupported!` right
there!

Also, this requires extending the `FromSql` impl for `DfType` to handle
`CITEXT` values, which is pretty straightforward

Refs: ENG-1893
Change-Id: I40186692fb29ca7f4936953c9f7024c3261024c3
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3367
Tested-by: Buildkite CI
Reviewed-by: Vlad Krasnov <vlad@readyset.io>
Reviewed-by: Nick Marino <nick@readyset.io>