replicators: Replicate CITEXT arrays
In the case where we have to manually construct a postgres_types::Type
for a snapshotted value from PG, we also have to explicitly load the
metadata of any array member types to correctly construct the `Kind`
containing the inner type (this eventually gets passed to the FromSql
impl for DfValue). This makes CITEXT arrays replicate, but also unlocks
replicating arrays of other "non-builtin" types such as enums.

Fixes: ENG-1913
Change-Id: Ib4d0622475c511b09dbf7fed1ea9e9dc7559f6a5
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3370
Tested-by: Buildkite CI
Reviewed-by: Vlad Krasnov <vlad@readyset.io>