dataflow-expression: Break builtins out into a module
Since the bulk of the complexity of evaluating dataflow expressions is
built-in functions (and this will only get worse as we add more and more
support for builtins!), it makes sense to separate them out into their
own module (which I've made into a submodule of `eval`). This also paves
the road for making *more* submodules of eval in the future, perhaps to
group together the implementation of builtins that deal with similar
types (such as time).

Change-Id: Ib77b0d88bbac61b5d62ceb7d8528d7d96daaa312
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3377
Reviewed-by: Nick Marino <nick@readyset.io>
Reviewed-by: Nikolai Vazquez <nikolai@readyset.io>
Tested-by: Buildkite CI