Store view in NoriaConnector::get_view()
In NoriaConnector::get_view(), we currently request but do not store a
View when we receive a query that the adapter has not seen before. Then,
the View is requested again when we subsequently call get_noria_view().

This change stores the View in get_view() so that the subsequent call to
get_noria_view() pulls the view from the cache instead of making another
RPC to the controller.

This change is not significant as it only affects our behavior for each
new query seen by the adapter, which impacts a short period at startup.

Change-Id: Ie4e36fe2e60914b7fe444c957bea4acb33ae3411
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3379
Tested-by: Buildkite CI
Reviewed-by: Peter Farr <peter@readyset.io>