Add `->` and `->>` operators to extract from JSON for PostgreSQL
These operators extract a JSON/JSONB value using an array index integer
or object key string.

The naming convention used for the dataflow enum variants is:
"Json", key type, operation, output type (if not JSON).

While parsing these operators is now possible under MySQL, using them
will result in an unsupported error.

This commit is dependent on being able to know whether we're working
in the context of MySQL or PostgreSQL. The MySQL version of this
operator is blocked on JSON path support.

Change-Id: Ic498d1b0e8238f34b264e679668a82139cdc3874
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3403
Reviewed-by: Nick Marino <nick@readyset.io>
Reviewed-by: Griffin Smith <griffin@readyset.io>
Tested-by: Buildkite CI