nom-sql: add position information to parsing errors
This commit leverages the nom_locate crate to provide line and column
numbers to SQL parsing errors. It does this by wrapping all input
sequences in nom_locate::LocatedSpan, and implementing a custom error
type for nom::IResult that propagates the error returned by the parsing
operation that made it the furthest, e.g. out of the branches of an alt
combinator.

The error output used to be:
- the slice of the input sequence that failed to parse
- the error "tag" i.e. the nom combinator that failed

Now, it additionally contains:
- the byte offset at which the unparsed input sequence begins within the
  total input sequence
- the line # of the same (and by computation, the column #)

As of this change, no error information gets propagated to the user,
only the message "failed to parse query". Doing so will be a follow-on
change. The error information makes it as far as the top-level parsing
functions e.g. nom_sql::parser::parse_query_bytes before being dropped
on the floor.

Co-authored-by: Dan Wilbanks <dan@readyset.io>
Change-Id: I00ee0e8c14a322197d4a3e4a0aef16f336396830
Refs: ENG-1923
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3473
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Reviewed-by: Luke Osborne <luke.o@readyset.io>
Tested-by: Buildkite CI