data: Define a separate Dialect type for expr evaluation
The notions of how strings get parsed into a SQL AST and how expressions
within that SQL AST get evaluated are distinct concerns - but up to this
point we've been conflating those two concerns under the umbrella of the
`nom_sql::Dialect` enum. This is not only conceptually a little fraught,
but also has genuine issues down the road once *both* the parsing
specifics and the evaluation semantics become configurable at runtime
(via things like MySQL's `sql_mode` flag).

This commit divides the two notions, defining a new
`readyset_data::Dialect` type (also re-exported as
`dataflow_expression::Dialect`) which currently looks a lot like
`nom_sql::Dialect`, but paves the road for us becoming a lot clearer
about which is which, and when to use each and in what context.

Change-Id: Id566aec99a2295859d0c3d21c663418eaecbb372
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3390
Tested-by: Buildkite CI
Reviewed-by: Nikolai Vazquez <nikolai@readyset.io>