Adapter: Determine pagination support at startup
This moves the pagination check from initation of a client
connection to startup of the adapter itself. This removes a failure
point which (if the check fails, e.g. excessive adapter load), from
resulting in a client connection hanging confusingly (i.e. the user
doesn't get any errors, just a hang).

Checking at startup is consistent with the ReadySet-server, as its
support for Pagination is determined at startup via CLI arguments.

As a (minor) side benefit, this should result in client connections
initializing a bit faster, as it removes a Adapter -> Server RPC
round-trip.

Refs: ENG-1700
Change-Id: I48c97045ca317d404c9462f3edaeeba18cf6b027
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3392
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke.o@readyset.io>