Add a `dialect` field to ChangeList
Add a new `dialect: Dialect` field to the ChangeList. This will
eventually be set based on the database engine we're emulating in the
adapter and threaded through to migrations to control evaluation and
type semantics of expressions, but for now it's hardcoded to the MYSQL
dialect in the adapter and ignored in the server.

The bulk of this change comes from the fact that we are no longer able
to directly parse a ChangeList from a string containing SQL, so there's
a lot of places, especially in tests, where we need to shuffle around
the API for that to pass in a Dialect. That's all mostly mechanical
though - the interesting bits here are all in `changelist.rs`

Refs: ENG-1418
Change-Id: Ib4b610cafb9d8a0f66b4fadcd8633172a7bd9462
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3394
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>