view: Use DfType as the type for column schema
Rather than (lossily!) converting the type for key and returned columns,
which we already have as DfType, to SqlType - only to convert
them *back* to DfType in a few places! - just return them as DfType
directly. This pushes the frontier for DfType one level forward, instead
opting (for now!) to convert to SqlType in the actual engine-specific
adapter binaries (albeit this time with the correct dialect, rather than
a hardcoded one!), right before we then convert to the protocol-specific
metadata type.

This takes the opportunity to make the structure of the `ColumnSchema`
struct a little bit nicer, since we have to change it anyway - now,
instead of embedding the AST struct for the list item in the CREATE
TABLE statement (!!!) we only include the actual fields we care about -
`comment` was always ignored, and `constraints` is used, but only set if
the column comes directly from a column in a base table, so we can
safely put it in `ColumnBase`.

Fixes: ENG-1848
Change-Id: I9fc1a4db61a744ea7d0f76045d73d06ca3e60be0
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3395
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>