Implement `TryFrom<[&]DfValue>` for all integers
This commit consolidates the implementation code into a macro, ensuring
the implementations are consistent with one another and reducing the
maintenance burden.

The addition of the `usize` implementation enables us to index into
slices using `DfValue`.

This also changes the logging behavior to mark the out-of-bounds integer
value as sensitive data.

For infallible conversions (e.g. `TryFrom<u64>` for `u128`), the error
conversion branch is never generated because `std::convert::Infallible`
is an uninhabited bottom type.

Change-Id: I503a7442a79a5f529d822e190fd4daaa615b89fb
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3402
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>