server: Propagate dialect through migrations
Add a `dialect` field to Migration set via the `Dialect` supplied to the
ChangeList, and propagate it to all the places we were previously
hardcoding dialect, including query rewriting, expression lowering, and
type conversion for base node creation.

There's *one* last place we're hardcoding dialect here, which is the
`migrate` method on `ControllerHandle` - since that's only used for
tests, but it's used a *ton* for tests, I figured it'd be best to leave
it hardcoded for now to avoid a massive change to the readyset-server
integration tests.

Refs: ENG-1418
Change-Id: I96c4c112ce58d5550aedbff2ee24c71f8a4fca1e
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3413
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>