server: Pass in a dialect to view_statuses
Since view_statuses rewrites queries, it's going to need to know the SQL
dialect to use for those queries once we stop hardcoding that. This adds
it as a flag to the RPC and threads it through the ViewsSynchronizer so
it gets passed in, but leaves it unspecified in the actual rewrite for
now in preparation for removing the dialect hardcoding everywhere.

Refs: ENG-1418
Change-Id: Iecf070f79be84846c681b0ca855cf5d1a86a7b0b
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3412
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>