readers: use notification for blocking reads
Currently blocking reads will retry every fixed period of time.
This check is very wasteful, and moreover tokio::sleep is also
surprisingly wasteful and consumes a lot of CPU. In addition the
retrigger for long-standing reads is also time based.

This change introduces a new notification based system for
blocking reads, where the writer notifies all existing readers
if a recheck is desirable. This makes sure readers only do the
recheck when the writer changes. This both reduces the idle
recheck CPU usage to 0, and eliminates the extra read delay for
missed reads, as the read is performed as soon as possible.

Moreover this change also skips retrigger if we know there was
no eviction since the last trigger, which eliminates the overhead
of replaying the same upqueries over and over for long standing
reads.

Change-Id: Id0a382626f79e5dec814d359a505b7d593775f3e
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3421
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke.o@readyset.io>
Reviewed-by: Dan Wilbanks <dan@readyset.io>