Inlining: Add StripLiterals SQL pass
Adds a pass to remove and return all literals from a query. The stripped
query and literals will be stored in Recipe to allow us to match
parameterized queries to inlined migrations.

Also adds a SelectStatementSkeleton wrapper around SelectStatement to
enforce that a particular SelectStatement has had the
StripLiteralsVisitor pass run on it.

This pass is inefficient since it will require us to clone the original
query. This could be avoided with a custom hash implementation, so that
we can hash the query without modifying it.

Change-Id: I7f7b73af210ff0b436521fade4205eae7500c50f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3446
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>
Reviewed-by: Griffin Smith <griffin@readyset.io>