Tweak logging around dry run failures
We attempt applying recipes during a dry run for up to
max_processing_minutes. During that time, the recipe may fail to apply.
Before this change, we were periodically emitting ERROR logs when we
failed to apply the recipe. Since such failures are expected to happen
occasionally, this made those error logs noisy.lSo this changes them to
INFO logs, with an ERROR log emitted if we are still failing after the
max_processing_minutes has elapsed.

It also logs an anonymized version of the query that failed to apply if
we hit the error condition.

Change-Id: Id9d00700ba7e33df504c9fe2583d6078c29ed0bb
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3420
Tested-by: Buildkite CI
Reviewed-by: Peter Farr <peter@readyset.io>