nom-sql: Support parsing ADD CONSTRAINT ... KEY
This change adds support for parsing ALTER TABLE ADD CONSTRAINT ... KEY,
where ... represents the different types of possible TableKeys. Prior to
this, only FOREIGN KEY allowed the optional 'ADD CONSTRAINT' prefix.

To support this, this adds a constraint_name field to most of the
TableKey enum variants.

To aid in debugging in implementing this, I added a `debug_print`
function that is a wrapper for `eprintln!()` if the debug feature is
enabled, or a no-op if it is not. This allows us to commit debug strings
and not have to add them back in each time we go to debug things in
nom-sql.

Change-Id: I7a4c6b42c499db0384cb27502bf3fa82e3d86121
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3425
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>