Expand fallback cache to support prep/exec
This expands the fallback cache to include a very naive way of
supporting prepared statements. This allows us to write benchmarks using
the system-benchmarks crate, which rely exclusively on prepared
statements.

The way this is *very* naively done is by taking a tuple of the prepared
statement id and the parameters, hashing it, and using the hash string
in place of the query string (key) in the fallback cache underlying map.

In principle this works but is very hacky.

Change-Id: I2d3a8607653333256fd5b307d6c4a388dec93b83
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3454
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke.o@readyset.io>