query_graph: Eliminate clones and vec allocs for split_conjuctions
This change modifies split_conjuctions to use iterators instead of
vectors, which lets us avoid a bunch of unnecessary vector allocations
and clones.

I don't know if this change will make a significant difference in
practice, but I wanted to try writing the code as an exercise for
myself, and then figured I might as well commit it since I got it to
work. I'm also planning to add an additional use of split_conjunctions
in a later commit as part of my work on implementing HAVING support.

Change-Id: If26bcc4d802d75419ccc8690946e2b3561d6b2f8
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3458
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>
Reviewed-by: Dan Wilbanks <dan@readyset.io>