query_graph: Add code to extract HAVING aggregates
This commit adds the new `extract_having_aggregates` function, which
walks the HAVING expression and pulls out aggregate functions, replacing
them with column references and storing the aggregate functions
separately in the existing `aggregates` map.

By storing the aggregates in the existing map, we can rely on the code
that's already in place to create the necessary aggregate nodes.

Note that we still do not actually create the necessary predicate nodes
to make HAVING clauses work. That functionality will be added in the
next commit.

Change-Id: I4be39b4040023af80ad49e12bc4e352754acf7e5
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3471
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>
Reviewed-by: Dan Wilbanks <dan@readyset.io>