client: Don't try to interpret `USE` via transaction boundary
If an upstream was present, we'd previously incorrectly try to send a
`USE` statement to `handle_transaction_boundaries`, even after we had
already handled it by setting our schema search path. This cleans up the
code paths there a *bit*, avoids sending the wrong statements to the
wrong methods, and covers USE statements when an upstream is present
with a test.

Change-Id: I2b3687d4e6bd2e87971f74adcb420bf136937a6a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3475
Reviewed-by: Luke Osborne <luke.o@readyset.io>
Tested-by: Buildkite CI