readyset-sql-passes: Add visit_sql_identifier
This adds visit_sql_identifier and uses it to provide more coverage for
what we walk. With that extra coverage, we can significantly reduce what
we need to implement for the Anonymizer visitor, since most things
eventually find their way to visit_sql_identifier.

This also fixes a bug where the field expr name wasn't properly being
anonymized, as it (a sql identifier) wasn't being visited.

Change-Id: I4fca3b0e7a6d866237a51711a320a3338259954a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3478
Tested-by: Buildkite CI
Reviewed-by: Alex Martin <Alex.m@readyset.io>