dataflow-expression: Implement DATE_FORMAT
Fully implement the MySQL DATE_FORMAT function for datetime arguments.
Since the format string rules that MySQL supports have very little
overlap with the time formatting that eg chrono supports, this is all
done entirely manually, offloading to chrono when we *can* but mostly
implementing stuff ourselves. By far the worst part of this is the
`week_and_year` function, which implements a nearly completely
inscrutable function for calculating the week number of a date, taking
into account the starting day of the week (monday or sunday) and
the *year* of the week (some days within a year are actually within the
last week of the previous year, or the first week of the next year).

Fixes: ENG-1893
Change-Id: I85de6c12acabbe6c428294d9bb3195d46304f69d
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3481
Reviewed-by: Nick Marino <nick@readyset.io>
Tested-by: Buildkite CI