dataflow: Inform nodes when their siblings are remapped
All Ingredient impls in Dataflow have fields set on construction that
store the node indices of their siblings (currently always parents,
though hypothetically that could stop being the case at some point in
the future). These node index fields were previously assumed to always
be completely static, but with the remapping introduced in
8d48e9add (Fix Partial Materialization Bug, 2021-10-13) it became
possible for a node's parents to change *after* it was created - though
we never actually told the node that this was the case. This commit adds
a new method to Ingredient called `replace_sibling` which is implemented
in all Ingredient impls (via a macro in all cases except Union) to swap
out a single node index in internal state with a remapped node index,
and then calls that method from migration when we do the remapping.

Refs: ENG-1713
Change-Id: I73d5a39091efd39f1c7d4b78565bd5d9be6bfa45
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3487
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>